separated PingPong into its own class
